requirements.txt
flask==2.3.2
flask-socketio==5.3.4
flask-session
openai==1.3.0
langchain
langchain-openai
langchain-community
langchain-core
chromadb
gTTS
whisper==1.1.10
beautifulsoup4
requests
numpy==1.24.3
torch==2.0.1
python-dotenv
pydantic
eventlet==0.33.3
python-socketio==5.8.0